financial_audit_env/server/requirements.txt
# Server dependencies for Docker
# These are the minimal dependencies needed to run the FastAPI server.
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
openai>=1.0.0
openenv-core>=0.2.0
